Move around some classes
