+++ src/main/scala/com/nec/spark/planning/aggregation/VeFinalAggregate.scala
@@ -1,15 +1,12 @@
 package com.nec.spark.planning.aggregation
 
 import com.nec.spark.SparkCycloneExecutorPlugin.source
-import com.nec.spark.planning.OneStageEvaluationPlan.VeFunction
-import com.nec.spark.planning.{PlanCallsVeFunction, SupportsVeColBatch}
+import com.nec.spark.planning.{PlanCallsVeFunction, SupportsVeColBatch, VeFunction}
 import com.nec.ve.VeColBatch
 import org.apache.spark.internal.Logging
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.catalyst.expressions.{Attribute, NamedExpression}
 import org.apache.spark.sql.execution.{SparkPlan, UnaryExecNode}
-
-import java.nio.file.Paths
 
 case class VeFinalAggregate(
   expectedOutputs: Seq[NamedExpression],
