+++ src/main/scala/com/nec/spark/planning/plans/OneStageEvaluationPlan.scala
@@ -17,13 +17,11 @@
  * limitations under the License.
  *
  */
-package com.nec.spark.planning
+package com.nec.spark.planning.plans
 
 import com.nec.cmake.{ScalaTcpDebug, Spanner}
 import com.nec.spark.SparkCycloneExecutorPlugin.{source, veProcess}
-import com.nec.spark.agile.CFunctionGeneration.VeType
-import com.nec.spark.planning.OneStageEvaluationPlan.VeFunction
-import com.nec.spark.planning.OneStageEvaluationPlan.VeFunction.VeFunctionStatus
+import com.nec.spark.planning.{PlanCallsVeFunction, SupportsVeColBatch, Tracer, VeFunction}
 import com.nec.ve.VeColBatch
 import com.nec.ve.VeKernelCompiler.VeCompilerConfig
 import com.typesafe.scalalogging.LazyLogging
@@ -34,36 +32,6 @@ import org.apache.spark.sql.catalyst.plans.physical.Partitioning
 import org.apache.spark.sql.execution.{SparkPlan, UnaryExecNode}
 
 import scala.language.dynamics
-
-object OneStageEvaluationPlan {
-
-  object VeFunction {
-    sealed trait VeFunctionStatus
-    object VeFunctionStatus {
-      final case class SourceCode(sourceCode: String) extends VeFunctionStatus {
-        override def toString: String = super.toString.take(25)
-      }
-      final case class Compiled(libraryPath: String) extends VeFunctionStatus
-    }
-  }
-
-  final case class VeFunction(
-    veFunctionStatus: VeFunctionStatus,
-    functionName: String,
-    results: List[VeType]
-  ) {
-    def isCompiled: Boolean = veFunctionStatus match {
-      case VeFunctionStatus.SourceCode(_) => false
-      case VeFunctionStatus.Compiled(_)   => true
-    }
-    def libraryPath: String = veFunctionStatus match {
-      case VeFunctionStatus.SourceCode(path) =>
-        sys.error(s"Library was not compiled; expected to build from ${path
-          .take(10)} (${path.hashCode})... Does your plan extend ${classOf[PlanCallsVeFunction]}?")
-      case VeFunctionStatus.Compiled(libraryPath) => libraryPath
-    }
-  }
-}
 
 final case class OneStageEvaluationPlan(
   outputExpressions: Seq[NamedExpression],
