+++ src/main/scala/com/nec/ve/VeRDD.scala
@@ -3,7 +3,7 @@ package com.nec.ve
 import com.nec.spark.SparkCycloneExecutorPlugin.source
 import com.nec.ve.VeColBatch.VeColVector
 import com.typesafe.scalalogging.LazyLogging
-import org.apache.spark.HashPartitioner
+import org.apache.spark.{HashPartitioner, TaskContext}
 import org.apache.spark.rdd.{RDD, ShuffledRDD}
 
 import scala.reflect.ClassTag
