+++ src/test/scala/com/nec/cmake/DynamicCSqlExpressionEvaluationSpec.scala
@@ -22,16 +22,9 @@ package com.nec.cmake
 import com.eed3si9n.expecty.Expecty.expect
 import com.nec.native.NativeEvaluator.CNativeEvaluator
 import com.nec.spark.SparkAdditions
-import com.nec.spark.planning.{OneStageEvaluationPlan, VERewriteStrategy}
+import com.nec.spark.planning.VERewriteStrategy
 import com.nec.testing.SampleSource
-import com.nec.testing.SampleSource.{
-  makeCsvNumsMultiColumn,
-  makeCsvNumsMultiColumnJoin,
-  SampleColA,
-  SampleColB,
-  SampleColC,
-  SampleColD
-}
+import com.nec.testing.SampleSource.{SampleColA, SampleColB, SampleColC, SampleColD, makeCsvNumsMultiColumn, makeCsvNumsMultiColumnJoin}
 import com.nec.testing.Testing.DataSize.SanityCheckSize
 import com.typesafe.scalalogging.LazyLogging
 import org.scalatest.{BeforeAndAfter, BeforeAndAfterAll}
@@ -40,12 +33,13 @@ import org.scalatest.matchers.must.Matchers
 import org.scalatest.matchers.should.Matchers.convertToAnyShouldWrapper
 import org.apache.spark.sql.internal.SQLConf.CODEGEN_FALLBACK
 import org.apache.spark.sql.{Dataset, SparkSession}
-import org.scalactic.{source, Prettifier}
+import org.scalactic.{Prettifier, source}
 
 import java.time.Instant
 import java.time.temporal.TemporalUnit
 import scala.math.Ordered.orderingToOrdered
 import com.nec.spark.planning.VeColumnarRule
+import com.nec.spark.planning.plans.OneStageEvaluationPlan
 
 abstract class DynamicCSqlExpressionEvaluationSpec
   extends AnyFreeSpec
